+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.3.10+1
+- fixed android build
+
+# 0.3.10
+- add mediaPlaybackRequiresUserGesture parameter
+- Add ignore ssl error parameter
+
 # 0.3.9+1
 
 - Fixed error methods on iOS